backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
opencv-python==4.9.0.80
numpy==1.26.4
face_recognition==1.3.0
dlib==19.24.2
SQLAlchemy==2.0.28
PyMySQL==1.1.0
bcrypt==4.1.2
PyJWT==2.8.0
web3==6.15.1
requests==2.31.0
urllib3==2.2.1
cmake==3.28.3
